+++ index.js
@@ -1,45 +1,54 @@
 import { NativeModules, NetInfo, Platform } from 'react-native';
-import ReactNativeFS from 'react-native-fs';
+import { deleteFile, download, isFileExists } from './lib/downloadHelper';
 
-const { RNAppUpgrade } = NativeModules;
+const { RNEasyUpgrade } = NativeModules;
 
 let jobId = -1;
+const isAndroid = Platform.OS === 'android';
+
+const VERSION_NAME = RNEasyUpgrade.versionName;
+const VERSION_CODE = RNEasyUpgrade.versionCode;
+const DEFAULT_DOWNLOAD_APK_NAME = `newVersion.apk`;
+
+const DocumentDirectoryPath = isAndroid && RNEasyUpgrade.RNDocumentDirectoryPath;
 
 const defaults = {
-  iosAppLookupUrl: 'https://itunes.apple.com/lookup?id={0}'
+  iOSAppLookupUrl: 'https://itunes.apple.com/lookup?id={0}',
+  downloadTitle: '下载更新包',
+  downloadDescription: '新版本更新包下载中',
+  downloadDestDirectory: DocumentDirectoryPath,
+  downloadApkName: DEFAULT_DOWNLOAD_APK_NAME,
+  downloadApkEnd: path => RNEasyUpgrade.installApk(path),
+  shouldCheckApkHasDownloaded: true,
+  onError: () => {}
 };
 
 class AppUpgrade {
   /**
    *
    * @param {object} options
-   * @param {string} [options.iosAppLookupUrl] 可选，ios检查更新的地址
-   * @param {string} options.iosAppId IOS的AppID
+   * @param {string} [options.iOSAppLookupUrl] 可选，ios检查更新的地址
+   * @param {string} options.iOSAppId IOS的AppID
    */
   constructor(options) {
     this.options = Object.assign({}, defaults, options);
   }
 
-  get defaultApkSavePath() {
-    return `${ReactNativeFS.DocumentDirectoryPath}/Temp_App${RNAppUpgrade.versionName}.apk`;
+  get downloadDestPath() {
+    return this.options.downloadDestDirectory + '/' + this.options.downloadApkName;
+  }
+
+  get downloadDestDirectory() {
+    return this.options.downloadDestDirectory;
   }
 
   get downloading() {
     return jobId > -1;
   }
 
-  progress(data) {
-    const percentage = (100 * data.bytesWritten / data.contentLength) | 0;
-    this.options.downloadApkProgress && this.options.downloadApkProgress(percentage);
-  }
-
-  begin() {
-    this.options.downloadApkStart && this.options.downloadApkStart();
-  }
-
   _handleError(err) {
-    console.log('downloadApkError', err);
-    this.options.onError && this.options.onError(err);
+    console.log('EasyUpgradeError', err);
+    this.options.onError(err);
   }
 
   getNetworkStatus() {
@@ -49,67 +58,84 @@ class AppUpgrade {
   }
 
   getLocalVersionInfo() {
-    const versionName = RNAppUpgrade.versionName;
-    const versionCode = RNAppUpgrade.versionCode;
-    return { versionName, versionCode };
+    return { VERSION_NAME, VERSION_CODE };
+  }
+
+  checkApkHasDownloaded = async (path = this.downloadDestPath) => {
+    return await isFileExists(path);
+  }
+  /**
+   * download file
+   * @param fileUrl
+   * @param downloadConf
+   */
+  downloadFile(fileUrl, downloadConf = {}) {
+    jobId = 1;
+    download(fileUrl, downloadConf)
+      .then(async res => {
+        jobId = -1;
+        if (this.options.shouldCheckApkHasDownloaded) {
+          await RNEasyUpgrade.moveFile(downloadConf.tempDownloadPath, this.downloadDestPath);
+        }
+        this.options.downloadApkEnd(this.downloadDestPath);
+      })
+      .catch(err => {
+        jobId = -1;
+        this._handleError(err);
+      });
   }
 
   /**
    * update app and install
    * @param apkUrl
    */
-  updateAndroidApp(apkUrl) {
+  async updateAndroidApp(apkUrl) {
     if (this.downloading) {
       return;
     }
 
-    const progressDivider = 1;
-    const downloadDestPath =
-      this.options.downloadApkSavePath ||
-      `${ReactNativeFS.DocumentDirectoryPath}/Temp_App${RNAppUpgrade.versionName}.apk`;
-
-    const ret = ReactNativeFS.downloadFile({
-      fromUrl: apkUrl,
-      toFile: downloadDestPath,
-      begin: this.begin,
-      progress: this.progress,
-      background: true,
-      progressDivider
-    });
-
-    jobId = ret.jobId;
-
-    ret.promise
-      .then(res => {
-        // console.log('downloadApkEnd');
-        this.options.downloadApkEnd &&
-          this.options.downloadApkEnd(needInstall => {
-            if (needInstall) {
-              RNAppUpgrade.installApk(downloadDestPath);
-            }
-          });
-
-        jobId = -1;
-      })
-      .catch(err => {
+    const tempDownloadApkName = 'temp_download.apk';
+    const tempDownloadPath = this.downloadDestDirectory + '/' + tempDownloadApkName;
+    const downloadConf = {
+      downloadTitle: this.options.downloadTitle,
+      downloadDescription: this.options.downloadDescription,
+      saveAsName: this.options.shouldCheckApkHasDownloaded ? tempDownloadApkName : this.options.downloadApkName,
+      allowedInRoaming: true,
+      allowedInMetered: true,
+      showInDownloads: true,
+      external: true, //when false basically means use the default Download path
+      path: this.downloadDestDirectory, //if "external" is true then use this path
+      tempDownloadPath: tempDownloadPath
+    };
+    if (this.options.shouldCheckApkHasDownloaded) {
+      try {
+        const isTempFileExisted = await isFileExists(tempDownloadPath);
+        if (isTempFileExisted) {
+          await deleteFile(tempDownloadPath);
+        }
+      } catch (err) {
         this._handleError(err);
+      }
+    }
+    this.downloadFile(apkUrl, downloadConf);
+  }
 
-        jobId = -1;
-      });
+  installApk(apkPath = this.downloadDestPath) {
+    RNEasyUpgrade.installApk(apkPath);
   }
 
   /**
    * Check the ios app version info from app store.
    */
   async checkAppVersionIOS() {
-    if (!this.options.iosAppId) {
-      throw new Error('Must provide iosAppId.');
+    if (!this.options.iOSAppId) {
+      throw new Error('Must provide iOSAppId.');
     }
-    const iosCheckVersionUrl = (this.options.iosAppLookupUrl || '').replace(/\{0\}/g, this.options.iosAppId);
+    const iosCheckVersionUrl = (this.options.iOSAppLookupUrl || '').replace(/\{0\}/g, this.options.iOSAppId);
     const data = await fetch(iosCheckVersionUrl).then(res => res.json());
     const result = ((data || {}).results || [])[0] || {};
     const latestVersion = result.version;
-    const localVersion = RNAppUpgrade.versionName;
+    const localVersion = VERSION_NAME;
     this.trackViewUrl = result.trackViewUrl;
     const hasNewVersion = latestVersion !== localVersion; // 简单认为，版本号不一致就是有新版本
     return {
@@ -126,7 +152,25 @@ class AppUpgrade {
    * @param {string} [trackViewUrl] [Optional] The app download page
    */
   navigateToAppStore(trackViewUrl) {
-    RNAppUpgrade.openURL(trackViewUrl || this.trackViewUrl);
+    RNEasyUpgrade.openURL(trackViewUrl || this.trackViewUrl);
+  }
+
+  startAppUpdate(apkUrl, appStoreUrl = this.trackViewUrl) {
+    if (isAndroid) {
+      if (this.options.shouldCheckApkHasDownloaded) {
+        this.checkApkHasDownloaded().then(async hasDownloaded => {
+          if (hasDownloaded) {
+            this.options.downloadApkEnd(this.downloadDestPath);
+          } else {
+            this.updateAndroidApp(apkUrl);
+          }
+        });
+      } else {
+        this.updateAndroidApp(apkUrl);
+      }
+    } else {
+      this.navigateToAppStore(appStoreUrl);
+    }
   }
 }