+++ CHANGELOG.md
@@ -1,4 +1,40 @@
-## 0.0.1 (2018-05-31)
+## 0.1.0 (2018-08-23)
+
+* add fileProvider in the project
+
+## 0.0.9 (2019-05-11)
+
+* add fileProvider sample in demo and fix bug
+
+## 0.0.8 (2018-12-27)
+
+* npm publish the new package
+
+## 0.0.7 (2018-09-04)
+
+* Fix RN 0.50+ android build failed bug and add docs&examples
+
+## 0.0.6 (2018-09-03)
+
+* Fix when tempFile download failed cannot install new apk bug
+
+## 0.0.5 (2018-08-17)
+
+* Add download temp cache check
+
+## 0.0.4 (2018-08-16)
+
+* Fix download apk path joint error
+
+## 0.0.3 (2018-08-14)
+
+* Fix that default download apk name is const value.
+
+## 0.0.2 (2018-08-11)
+
+* Remove `react-native-fs` package
+
+## 0.0.1 (2018-08-10)
 
 * placeholder on npm
 * no function